requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
motor==2.2.0
pydantic[email]
python-decouple==3.4